fix: restore non-root container support
Unlike the old version of non-root container support, all user and group
information in the tar archives destined for the container are
configured entirely on the master side, before submitting the container
specifications to the agent.  This is much cleaner, and removes the need
for the special -1 UID/GID value previously used to tell the agent to set the
UID/GID to whatever the container will run with.

This also fixes a new bug we had where tarballs stored in the database
would print errors like:

    tar: Extended header gid=-1 is out of range 0..4294967295

during the untar process.

In order to actually support the runtime `pip install` step, all runtime
pip installs now include the `--user` option.  In order for `--user` to
work, the PYTHONUSERBASE variable is now set to a subidr of
/run/determined in all containers, so that we can guarantee that the
user owns the directory `pip install --user` owns without making any
assumptions about the HOME environment variable.