refactor: change default storage dir permissions
In a world with non-root containers, the call to os.makedirs when the
StorageMangager detects that a storage path does not exist should create
the storage directory to be world writable, so that other containers
running under other users do not fail.  This only really applies to
shared_fs storage configurations.

Users with tighter security needs should set their storage permissions
beforehand, or alternately restrict the permissions at the host_path.