feat: track the context directory as part of train context.
Previously to this commit, the context directory was assumed to be the
current working directory in the PyTorchTrial controller implementation.
This commit moves it to be tracked explicitly as part of det.EnvContext.

Note that this has the side-effect of adding a context directory
argument to local-development related APIs. The context directory is set
to the current working directory if left unspecified in these APIs.