chore: lower max_connections of postgres instances
DET-2812 revealed determined-master was using too many cnxs.
The approach settled on was to limit the cnxs on the client side.
This change sets the instance's limit slightly above the client side to
account for admin/cloud/etc provider tools. For reference, in an
extended run, gcloud (from cloudsqladmin) maintained 5-10 cnxs.
det-deploy aws was left untouched because serverless is unconfigurable
and defaults to 180.