chore: lower task container ssh key size
When we have an experiment with thousands of trials, we generate an ssh
key for each trial. This becomes very expensive, since ssh keygen is a
cpu intensive operation. By reducing the keysize, the performance
generating the keys moves from ~1.4 s/op to ~0.02 s/op, which is enough
solve the present impact. Security concerns were taken into
consideration but the worry is negated by the key only being used for
a short duration - usually much less than a day. The time to crack
an 1024-bit RSA key is much longer given even extraordinary
circumstances.