docs: update `CONTRIBUTING.md`
Many parts of `CONTRIBUTING.md` were quite out-of-date; this updates
everything (more or less) to match what is correct now. More could
probably be done to make it nicer as a how-to-contribute doc
overall (not least the addition of some content in the "Contributing
Changes" section), but at least the egregiously wrong parts are gone.