fix: add det command to PATH
Add the det command to the PATH in the entrypoint scripts.

It cannot be set with other environment variables in the master because
the docker.Config object will not do environment variable expansion.

The alternative would be to somehow, in the master, inspect the metadata
for the docker container to get the PATH environment variable, then use
that to set the final PATH, but then you would need to have the image
present on the master for that to work.