fix: move a new SQL migration after all previous ones numerically
Due to a long-running PR, we landed a migration after one with a larger
version number. As a result, upgrading existing databases is broken,
since the new migration is being entirely skipped over by the migration
library. This moves that migration to have a higher number so that it
will be picked up and run as normal when a database schema contains the
older migration but not the newer one.

It also changes the name of the function introduced in the migration,
both making the name more correct and avoiding the need for manual
intervention in databases that have been freshly created since the
original change containing the new migration landed (which would
otherwise fail on upgrade due to a duplicate function name).

(It also adds a `public` schema qualifier to the down migration, which
won't change anything in practice but matches the existing migrations.)