feat: exit log tailing based on state and time rather than polls
Previously, the decision to exit a `det trial log -f` was made based on
the number of polls since any new messages had appeared; after every 10
polls that passed with no updates, the CLI would query the state of the
trial and exit if it was terminal. This commit changes two things:

- The trial state is now queried every time, so we know as soon as
  possible when it changes to a terminal state. (Since this is done in
  the same HTTP request and SQL query as fetching the logs, the cost
  should be minimal.)

- The final decision to exit is made based on time, rather than number
  of polls, since the last message and state change.

This means we get more accurate detection of trial termination while
being less vulnerable to a slow master causing the command to pause for
a long time after the trial finishes.