feat: make pending commands killable
Previously, killing a command in PENDING state (i.e., one for which no
containers had started running yet) caused it to receive a
`TaskTerminated` message, which it promptly dropped on the floor, since
termination logic was in the `ContainerStateChanged` handler. Here, we
introduce a new `TaskAborted` message to signal that case (handled by
the command actor but ignored by the other ones). It's a bit redundant
with the existing messages, but this is an early step toward
transitioning toward a different set of messages as part of a scheduler
refactor.