fix: add script to run inside agent's container
This change aims to address the issues with the
original startup script without removing the
functionality since it is also useful. The
problem with the original script is that it runs
outside of the environment that the agent runs.
Because of this things like configuring the
docker client don't work well. Alternatives to
this are supporting secrets for altering the
experiment docker config (similar to docker
login) or running the agent natively, both of
which we aren't quite ready for yet.