fix: hardcode shared_fs container path
The shared_fs.container path configuration variable becomes a Target for
a Docker bind mount. More specifically, it is the mount point of the
shared_fs checkpoint storage location inside a Determined workload
container. The container path was not parsable if the value
was set to anything other than "/determined_shared_fs". Since we do not
generally allow users to interact with the checkpoint directory during
trial execution we are removing the option to configure the container
path by hard-coding the value. We will later remove this value from the
configuration entirely.