docs: sort attributions by name
The Go ones are after the WebUI ones; otherwise, it looks a bit weird to
have a big chunk of "github.com/..." and so on with a few WebUI ones on
either side.

The sorting was accomplished by means of a modified version of rstfmt
(https://github.com/dzhu/rstfmt).

Also disable highlighting while we're at it, since otherwise the blocks
get highlighted as Python and have words like "any" and "if"
highlighted.