fix: add logic in agent to invoke docker credentials helpers
When we added in the container_startup_script to run in the
agent's runtime environment (docker), an assumption was made
that the golang client would handle invoking credentials
helpers. Unfortunately, they are invoked by the cli itself.
Because of this, this commit is to invoke the credential
helper if auth wasn't provided on the experiment config.