fix: include example name as prefix in example tarballs
We previously created tarballs for the examples without a directory
prefix. As a result, when extracted via `tar` or similar tools, files
would be placed in the shell's present working directory. That is
generally not considered good form.

Instead, we now prefix the files in the example tarballs with the name
of the example. For example, the contents of `mnist_pytorch.tgz` are now
nested under a top-level directory `mnist_pytorch`.