fix: prevent multi-gpu trials from hanging
Previously during multi-GPU training if a worker process died, if there was
no hvd.allreduce() call prior to the next barrier (e.g., if the error
occurred while reducing metrics) the worker would die but the trial would hang.
This would happen because the  SubProcessLauncher monitors the liveliness of the
horovodrun process on the chief machine and the liveliness of the sshd process
on the non-chief machines. These processes fail when a worker process fails
during or before a hvd.allreduce() call. Now, when an exception occurs in the
subprocess, we notify the subprocess launcher.

DET-2879  #Done.
DET-2882  #Done.