feat: be more lenient about errors in running `nvidia-smi`
When an agent Docker image is built with the Nvidia runtime and run
without it, the container winds up with an empty `/usr/bin/nvidia-smi`
file, which would cause the agent to get an unexpected error and fall
over. This changes the agent to log any such errors and continue running
instead. It is perhaps a bit broad to handle _all_ errors like so, but
it doesn't seem that helpful to add a bunch of code for ferreting out
very specific errors, and I think it does make sense to allow the agent
to be robust here.