feat: make the agent work with `nvidia-container-toolkit`
As of version 19.03, the recommended way of running Docker containers
with GPUs changed from `nvidia-docker2` to `nvidia-container-toolkit`.

That comes with a new way of configuring GPU containers:
rather than setting the runtime and `NVIDIA_VISIBLE_DEVICES` environment
variable when creating a container, we set a field for device
requests (new in Docker API version 1.40) during container creation.

Note that this adds a hard dependency on Docker 19.03, but we already
had one by default, since the Docker client library defaults to API
version 1.40 and no negotiation, which requires Docker 19.03.

Also, the new system makes it more annoying to check whether Docker has
GPU support enabled; rather than just looking in the list of runtimes,
the best available method seems to be to start a container with GPUs.
That takes a fair bit of code, requires looking for a particular error
from Docker, and should only actually help in a limited set of
circumstances (agent is running outside Docker, machine has GPUs, Docker
is not already set up to use GPUs) that can arguably be categorized as
missing prerequisites, so we just drop that check.