fix: exit out of non-chief containers gracefully
To exit out of a trial, the harness processes a Terminate message. When using a subprocess
to train (multi-GPU), the subrprocess launcher waits to receive a reply from all the subprocesses
or until a health check fails while a Terminate message is being process to reply to the socket and
workload manager that the Terminate has completed. Previously, the trial controllers would not respond
to the Terminate message. On the chief agent that started  horvodrun, the health check would fail because
the horovodrun process would exit out once all the subprocesses completed, and the containers would
terminate with error code of 0. On the other agents, the sshd process would remain alive, thus
the health checks would never fail, and the container would not exit until forcibly being killed,
resulting in a 137 error code. Now all processes respond to the terminate message, allowing all
containers to self-terminate.

DET-2945  #Done.