ci: allow manual approvals to be provided immediately
The manual approval request steps (for GPU tests and deploying) in
CircleCI had dependencies on the Docker build steps, which meant we had
to wait a few minutes after pushing to provide the approvals. Removing
the dependencies will mean we can immediately approve after pushing and
then let the build and tests/deployment run from there.