feat: add support for SAML
To meet customer needs, we need to provide support
for users to authenticate with SAML. The commit adds
/saml/sso and /saml/initiate to receive SAML responses
and initiate SP-initiated flow with SAML requests,
respectively. It also adds SSO providers to /info for
the frontend to determined if SAML is enabled or not and
the ability to configure SAML via master.yaml.