docs: make Sphinx default to not highlighting code blocks
Sphinx defaults to applying Python highlighting if possible [1], since
reST is so closely associated with Python. In our docs, though, we
actually have a pretty broad mix, so we don't want to end up with Python
highlighting where it doesn't apply.

This changes the default to no highlighting and removes all explicit
instances of "text" highlighting, which also means no highlighting.

[1] https://www.sphinx-doc.org/en/master/usage/restructuredtext/directives.html#code-examples