refactor: avoid dependending on pointer equality in scheduler test
We already define a notion of equality for TaskSummary, so using pointer
equality to define subset and difference for slices of TaskSummary is
confusing. Switching to TaskSummary.equals() is substantially slower,
but since this is only used in test code, that shouldn't be problematic.

AgentSummary didn't previously define an equality method but it is
trivial to define one -- so this commit does that and makes a
corresponding change for slices of AgentSummary.