refactor: remove dependency on GPUtil
Given the complexity of GPUtil and how little we need from it, it makes
sense to remove that dependency and get what we need from `nvidia-smi`
directly.

This code also provides better error handling in the face of strange
output from `nvidia-smi`.

Also fix up some type annotations; GPUtil returns IDs as integers, and
this code does the same, but the annotations thought they were strings.