feat: add container count to "det agent list"
When working with dynamic agents on a cluster running a variety of
workload types (e.g., experiments, notebooks, tensorboards, etc.), it
can be difficult to tell whether an agent is idle or not. This commit
adds a simple way to do that via "det agent list".

Note: this breaks CLI backward compatibility, as "det agent list" with
CLI >= 0.12.5 won't work with masters < 0.12.5.