feat: add model code to checkpoints
PyTorch and TF.1 Keras models need the users code to retrieve the model.
This commit adds the user code to all checkpoints.

TF.1 keras models can be saved using the tensorflow saved_model format
which doesn't require the original code to retrieve the model. This
would require running in eager execution mode for all trials which is
substantially slower than graph mode and thus not a viable option.