feat: support `--csv`, `--json` to `slot list` and `agent list` in CLI
This is useful as a stop-gap for folks that want to probe the state of
the cluster, e.g., to determine if there are free GPUs to run
low-priority batch inference jobs using the command infrastructure.

To make JSON output consistent with non-JSON output, we now construct an
intermediate list of dicts with the formatted version of the JSON we get
from the master.

Note that there are other `list` subcommands that might benefit from
being treated consistently, but this commit does not attempt to add
support for all of them.