refactor: remove container_path from configuration
container_path is known to the creator of the StorageManager or the
TensorboardManager, but should not be passed in as a configuration
entry; that should be a detail which is invisible to the user.

Shuffle unit tests of the SharedFSConfig validation from the harness
to the master, since validating at harness runtime is too late to be
a good user experience.