feat: support early trial termination
We now support trials erroring without closing the experiment. The
behavior for how this is left up to each search method, and is handled
strictly internal to the search method. This means that any mocked
metrics are not saved in the db or exposed to the user.

  - Random: No further action is needed.
  - Grid: No further action is needed.
  - ASHA-based: Treat the trials as having a MaxFloat64 validation
    metric. Mock promotions and ensure that no further workloads are sent
    to the experiment.
  - PBT: treat the trials as having a MaxFloat64 validation metric.
    Ensure that no further workloads are sent to the experiment, and we
    do not create new trials based on the completed workload.

As part of this work, we have introduced a flag to the shutdown message
that indicates to the experiment whether the searcher has completed
successfully. The current method that we use is we check that no trials
have exited early, which works for all search methods.