fix: avoid saving pytorch model architecture (#594)
In a previous iteration of checkpoint export, we used torch.save() to save the entire model architecture. We had to switch to only loading weights to make checkpoint export more robust.

This change removes the now unnecessary torch.save() which was failing with the Native API and custom PyTorch layers.