feat: checkpoint export from database fields (#664)
Checkpoint export writes out a `metadata.json` file to the checkpoint
directory during training storing the experiment configuration,
hyperparameters, and other relevant information (later used to load the
checkpoint). This file doesn't currently contain the checkpoint metadata
added as part of the model registry project. Additionally, the hparams
and experiment configuration are queryable at download time and do not
have to be written out to every checkpoint.

This PR will change the `Checkpoint.download` method to query the
database in order to construct the `metadata.json` file.

[DET-3272]