feat: add top level resource provider
The top level resource provider forward all requests to underlying
resource providers, for now we support just the default resource
provider. The top-level resource provider is now also responsible
for initializing the agents actor, which had previously been the responsibility
of the main actor.

As part of this change, scheduler messages types that are used by the
agents were moved out to "pkg/" to avoid circular dependencies.