feat: add a timeout to trial termination
When a trial terminates, if the process does not exit, the trial
container just sits there indefinitely; this can happen, for instance,
when model code spawns threads that don't terminate properly.

This change adds a timeout to trial termination: a fixed time after a
trial actor sends out a terminate message (whether the trial is
completely finished or just being paused), it sends a kill request to
the cluster.

Once a request for termination has been sent out, any task termination
errors are ignored, since the master has all it needs from the task;
whatever happens afterward is not really our concern.

Perhaps having a trial time out should count as an error, but it sounded
like hangs may be coming up in situations where it could be hard to fix
the hang (e.g., a third-party library spinning up threads and not taking
care of them properly) and it would be awkward to cause every trial to
fail in that case.