chore: refactor master to send batches in RUN_STEP
As part of the effort to remove steps from the UX, the master needs to
dictate to the harness the number of steps it should run per batch.
(And total seen so far). This enables the searcher to be configurated in
terms of records/batches/epochs and manage the conversion to steps
internally.