fix: upgrade scheme when using websockets
When we use websockets to connect to the master, historically we
have not changed the protocol to ws/wss. This works fine for ws
but in the case of wss, it causes lomond (the python library we
use) to perform the ws upgrade before a tls handshake is
completed---causing the connection to be rejected.