feat: support Pytorch multiple optimizers and LR schedulers [DET-3194, DET-3195, DET-3196, DET-3197, DET-3198]
In the new interface for this support, users need to initialize the
model, optimizers, and LR schedulers in the constructor of the trial
interface. They need to call the wrapper functions in the context object
to initialize these objects under distributed training settings.
They also need to initialize amp themselves.