fix: use custom TLS cert only for Determined API requests
Previously, if the CLI found a custom trusted master certificate, it
would be used for all requests made using Requests; that was the
simplest thing to write, but it meant that other HTTPS requests would
fail (e.g., checkpoint downloads from GCS, which go through Requests
internally). This change restricts the cert to only be applied to
requests made to the master.