feat: add top level resource provider [DET-3179, DET-3180] (#684)
As part of this change, we renamed the old `Cluster` to `DefaultRP` and added a top-level scheduler called `ResourceProviders` which handles all incoming messages and forwards them to the DefaultRP. In the future, this will allow us to add additional resource providers.

Agents actor is now initialized by the resource provider. This introduced a circular dependency between `internal/agent` and `internal/scheduler.` To handle this we moved the scheduler message types that are used by `internal/agent` to `internal/sproto`.