fix: fix nonroot dtrain and nonroot shell
Inject auxiliary passwd, shadow, and group files to augment the user
system inside the container.  This functionality depends on
libnss_determined being installed in the container.

This change also augments the command summary to include the
AgentUserGroup that each command is running with, and that information
is used by the CLI to choose the username when connecting to the sshd
server inside a shell container.

Finally, UsePAM was disabled in the sshd_config file, since the man page
explains that UsePAM is incompatible with running sshd as a non-root
user.  An alternate solution would be to run sshd as root but only
inject keys for the nonroot user, but that would have introduced the
following drawbacks:
  - It might not have satisfied some security-sensitive organizations.
  - It would not have worked for dtrain, which needs to run as non-root.
  - It would have been a larger divergence from the other command infra.