chore: upgrade agent VM image to newer kernel version
Upgrade the agent VM image to one built on top of a newer version of Ubuntu 16.04. This was motivated by a desire to mount a Lustre filesystem onto agent VMs, but the Lustre client requires a kernel version that is newer than what our base Ubuntu 16.04 image had.

The Docker contents did not change, but the change to the Environments repo changed the hash in the tag name, so we upgrade all references to the old hash to make the codebase consistent/easy to understand/easy to search.