docs: add docs on data access for dtrain [DET-3506] (#872)
Add documentation around data access to the Distributed Training How-To Guide. Briefly explain what ranks are and why data download directories should include the rank in their name. The motivation was that this is an important detail and was hidden in a subsection that distributed training users were unlikely to find.