chore: restore building individual tarballs for each example
When we split the examples into `native/` and `trial/`, this resulted in
building two giant tarballs, `native.tgz` and `trial.tgz`. Fix this so
that we build tarballs for each individual example again.

Note that this change assumes that there will not be naming conflicts
between native and trial examples.