feat: Add support for G4 instances on AWS [DET-1771]
We previously excluded G4s from the instance type whitelist; this commit
adds them to the whitelist and updates the documentation accordingly.

Note that the previous default agent AMI did not work with G4 instances.
The exact cause is unclear, but upgrading the Nvidia display driver from
410.104 to 450.51.05 appears to fix the problem (despite the fact that
Nvidia's docs claim that 410.104 supports T4 GPUs). The upshot is that
actually using G4 instances on AWS will require using an updated AMI.