feat: remove steps from config [DET-3298, DET-3299, DET-3525] (#937)
This change allows all searchers and min_validation_period and min_checkpoint_period to be configured in terms of records, batches or epochs. It also modifies all searchers to emit operations, such as Train{Epochs: 2} allowing searchers to be oblivious to how requested work is broken into workloads. Instead, the trial workload sequencer handles this and relays completed operations back to the searcher. It still relays workloads back to the experiment, though this is only to check the status of validations (is it the best?) and to record units complete so that progress is more granular.