feat: support task and agent summaries for k8 RP [DET-3424, DET-3425] (#891)
Support task summaries and agents endpoints. As part of this PR we do not attempt to change the structure of endpoints to be more suitable for k8s, instead we fit the k8s RP to work with existing endpoints. This means there is no support for agents/* endpoints.

To simplify integration with our actor system which routs api requests (for old api) by actor name, I added a forwarding mechanism which supports a preset actor name as the destination, instead of parsing the request path.