feat: remove steps from docs [DET-3626] (#939)
This change removes all usages of steps from the documentation, updating it to use the configs introduced by #937. Where steps was used as a training length, references were reworded as training length. Where usages of steps were used as the determined concept of a step, training workload was used in its place.