chore: update cloud images and default task envs.
This pulls in the recent update to the Nvidia drivers in the public
cloud images, moving from driver version 410.104 to 450.51.05.

Taskenvs should not actually be modified by this change, but since the
build machinery pushed a new version of the taskenvs, update to use the
latest taskenv versions just to be safe.