chore: remove steps from the CLI [DET-3625] (#940)
This change removes any references to the concept "step" in the CLI and replaces them with references to the number of batches complete at each workload. It also adds a database migration to backfill the values for `num_batches` and `total_batches_processed` for steps belonging to experiments trained in older versions of Determined.