docs: reformat experiment config documentation.
Replace the original giant bulleted list with a collection of sections
and sub-sections. This should hopefully make it easier to find content
within this documentation page, to link to the documentation for
specific configuration variables, and to include more discussion about
particular configuration settings when appropriate.

Along the way, add a few examples and fix/rewrite some of the docs for
different variables.