fix: use random port for notebooks [DET-3648] (#996)
Addresses the host mode networking issue where only a single notebook can start on an agent. This PR uses the same strategy as TensorBoards. We pick a random port per notebook within a range of ports. This is a mitigation. There is still a chance that 2 notebooks select the same port, although it is unlikely.