feat: remove steps from the UX (#968)
* feat: remove steps from config [DET-3298, DET-3299, DET-3525] (#937)

This change allows all searchers and min_validation_period and min_checkpoint_period to be configured in terms of records, batches or epochs. It also modifies all searchers to emit operations, such as Train{Epochs: 2} allowing searchers to be oblivious to how requested work is broken into workloads. Instead, the trial workload sequencer handles this and relays completed operations back to the searcher. It still relays workloads back to the experiment, though this is only to check the status of validations (is it the best?) and to record units complete so that progress is more granular.

* feat: remove steps from examples [DET-3624] (#938)

This change updates all examples to use the configuration changes made by #937. There was an effort made to convert the configurations to epochs where it was obvious; elsewhere, a direct conversion was made using `max_length.batches = batches_per_step * max_steps`.

* feat: remove steps from docs [DET-3626] (#939)

This change removes all usages of steps from the documentation, updating it to use the configs introduced by #937. Where steps was used as a training length, references were reworded as training length. Where usages of steps were used as the determined concept of a step, training workload was used in its place.

* chore: remove steps from the CLI [DET-3625] (#940)

This change removes any references to the concept "step" in the CLI and replaces them with references to the number of batches complete at each workload. It also adds a database migration to backfill the values for `num_batches` and `total_batches_processed` for steps belonging to experiments trained in older versions of Determined.

* feat: remove steps from preview-search [DET-3676] (#941)

This change updates the `det preview-search` functionality to work with the changes in #937 that removed steps from searchers.

* chore: rename batches_per_step [DET-3627] (#985)

This change removes the configuration setting batches_per_step from Determined and replaces it with scheduling_unit. It also changes a few instances where batches_per_step * step_id was used in SQL queries to determine the batch # at a step to prior_batches_processed + num_batches.=