feat: make agent's TLS verification level configurable
This enhances the agent's configuration regarding use of TLS on the
master connection to better support a range of deployment scenarios:

- Verification can be turned on or off.
- A custom CA certificate chain can be specified.
- TLS can be toggled separately from TLS for the agent's own API.

This technically makes the agent config/CLI options
backward-incompatble; part of the previous `--tls` option is now
splitting into `--security-tls-enabled`. However, it is unlikely that
anyone was using that, since the system cannot currently be run fully
over TLS anyway.