fix: handle failed validations correctly [DET-3838] (#1069)
Previously, a change was made that prevented completed validation workloads from completing in the experiment actor if they were missing validation metrics. This doesn't work because validation workloads that fail are intentionally missing metrics. Instead we check if the exit reason is nil before trying to compute the best validation.