fix: preserve rng state for all cloud access
Lots of cloud storage APIs use the built-in random package RNG to do
automatic random backoffs, which affects the random state that we try to
preserve for ML reproducibility.

Take the heavy-handed solution, and just preserve the state for every
single StorageManager or TensorboardManager call that might affect the
RNG state.