feat: add ability to perform initial validation [DET-3584] (#1061)
This change adds the ability to perform initial validations for trials. It also adds the concept of a zeroth step; it is special because it can be assumed to have a checkpoint, since we never train in a zeroth step. This simplifies rollback logic a lot - if anything is completed in a zeroth step (for now only validations), let it be (generally, we rollback to the last checkpoint).