chore: delete useless test (#1080)
test_dataset_restore tested functionality that we no longer have in the
system.  In reality, the test didn't even test that functionality when
it was written; it was a placeholder test for the right functionality
after a certain issue was addressed (DET-834, and arguably DET-836).

Because the test never really worked, nobody noticed when we removed the
dataset restoring functionality entirely.  But the test offers no real
coverage anymore.