feat: support user specified pod template specs [DET-3716, DET-3731] (#1005)
Support user specifying pod template specs in the experiment config. We validate that the user is submitting only supported fields. As part of this PR also addressed two bugs: 1) previously if an error occurred during pod creation, for trials we would fail in an infinite loop. 2) If a failure occurred while deleting Kubernetes resources (this always happens if the creation fails) we would not delete all the resources and fail to notify the task handler and the scheduler. Also as part of this PR added CI functionality to determine if running on k8s, enabling k8 specific coverage.