fix: handle failed checkpoints correctly in trial_workload_sequencer [DET-3853] (#1083)
This change fixes a bug in the trial workload sequencer where it is assumes that checkpoints won't fail. In the event of a failure checkpoint, a nil pointer is de-referenced when trying to access checkpoint metrics that don't exist in a failure. Now failures are checked before the completed message is handled. If it is an error, we count it as undone in the sequencer, except in the event it was a 'user_canceled' error message, which is counted as successful and forces a checkpoint after the current step.