fix: record progress in partial units (#1091)
This change alters trials to report units completed as a float64 so that partial units can be reported correctly. Previously, if the searcher was configured in terms of epochs and anything less than an epoch was completed in a workload, the reported completed model.Length was rounded to the nearest epoch; this usually ends up being 1 (except in the case of unusually small epochs or unusually large scheduling units) and makes progress appear completely wrong.