fix: let shell proxying work with master TLS
This change updates the tunneling helper script to allow proxying shell
connections to work when the master is listening on TLS. The main change
is a new HTTP connection class that applies TLS to the entire
connection; the standard library class makes the CONNECT request
unencrypted and only applies TLS to the proxied connection, which
doesn't make much sense for us.

To allow the copier thread to easily work with SSL sockets (which don't
have a useful fileno), normal sockets (which don't have read/write
methods), and stdin/stdout, there's also a class that wraps sockets to
provide read/write methods.