fix: proxy shell through master via HTTP CONNECT (#1067)
Add a CONNECT handler to our Echo framework.  The new handler is part of
the old proxy, which now supports proxying raw TCP trafic as well as
HTTP and websocket traffic.  The tunneled traffic is plaintext
naturally, but it is TLS-encrypted between master and client if
connecting to the master over TLS, and it is also end-to-end
ssh-encrypted for all `det shell` traffic.

Introduce a executable determined_cli.tunnel module, which will tunnel
its stdin and stdout through an HTTP CONNECT proxy, exactly the way
`nc -X CONNECT -x` would work, but without introducing an external
dependency on netcat.

Fix `det shell` to add a ProxyCommand="..." option to the ssh invocation
so that ssh calls the determined_cli.tunnel tool to tunnel traffic
though the master.

This fix allows shells to work in cloud environments.

Re-enable the old e2e test for determined shell, which is marked as a
GPU test so it will execute in a cloud context.

Add a new CPU e2e test for nonroot shell support.