fix: run at least one batch per searcher op (#1095)
This change fixes a bug where a very small adaptive searcher, in which the first rung is less than one batch, would end up requesting to run zero batches because it rounds down to the nearest batch. We now round down in the typical case, but round up if the searcher requested a training length less than the trial's global batch size.