chore: add a shell-entrypoint.sh
Add shell-entrypoint.sh which configures an interactive session with the
proper Determined python packages pre-installed and the proper PATH in
the same way that notebook-entrypoint.sh does.

Also configure ssh to set PermitUserEnvironment to yes, so that we can
use that mechansim to inject environment variables into the shell
sessions via the sshd server.

Also set PermitUserRC to yes and AllowEnv to * so that users can easily
customize their shell environment via some of the normal ssh mechanisms.