feat: kill idle tensorboards [DET-3808] (#1104)
Idle TensorBoard instances are spun down after a configurable amount of time (via master.yaml). Adds LastRequestedTime to the services registered in the proxy. The tensorboard manager calls itself on a tick. Every 5 seconds it checks for TensorBoard instances that haven't received traffic after the configured amount of time. These instances are killed.