feat: allow Python libraries in commands to use TLS for the master
This passes the same file and environment variables that we use for the
harness to command containers (which includes shells and notebooks (and
TensorBoards, though those won't actually use this)) to allow our Python
code within them to correctly connect to a master using TLS, including
the case of a custom certificate.

Having to handle the cert specially inside the container in the client
code is somewhat annoying, but I don't know of a way to install it
generally when we don't necessarily have root in the container.